scraper/requirements.txt
fastapi
uvicorn
firecrawl-py
python-dotenv
beautifulsoup4
requests